+++ mvc/src/main/java/com/interface21/webmvc/servlet/view/JsonView.java
@@ -1,7 +1,6 @@
 package com.interface21.webmvc.servlet.view;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.interface21.webmvc.servlet.View;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
